Refactor testing functions
Since those are general-purpose functions, move functions
testing normalization to mparray._testing
